Feat: GiftRepository Repository 처리에 대해서 재사용성을 할 Repository 가 없으므로, 실제 구현체로 네이밍 변경 및 비즈니스 로직 분리
